requirements.txt
opencv-python==4.9.0.80
pillow==10.2.0
matplotlib==3.9.0